--- a/src/query/sql/src/planner/plans/join.rs
+++ b/src/query/sql/src/planner/plans/join.rs
@@ -619,16 +619,6 @@ impl Operator for Join {
             ]);
         }
 
-        // (Serial, Serial)
-        children_required.push(vec![
-            RequiredProperty {
-                distribution: Distribution::Serial,
-            },
-            RequiredProperty {
-                distribution: Distribution::Serial,
-            },
-        ]);
-
         Ok(children_required)
     }
 }