… Cargo.lock
… tests/sqllogictests/suites/mode/cluster/memo/mix_property.test
… tests/sqllogictests/suites/mode/cluster/window.test