I hereby agree to the terms of the CLA available at: https://docs.databend.com/dev/policies/cla/

## Summary

`Serial` is bad for distribution execution, we should avoid it as much as possible.

Before
```
mysql> explain SELECT
    ->     e.name AS EmployeeName,
    ->     d.department_name AS DepartmentName,
    ->     e.salary AS Salary,
    ->     ROW_NUMBER() OVER (PARTITION BY d.department_id ORDER BY e.salary DESC) AS SalaryRank
    -> FROM
    ->     employees e
    -> JOIN
    ->     departments d ON e.department_id = d.department_id
    -> ORDER BY
    ->     DepartmentName,
    ->     SalaryRank;
+------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
| explain                                                                                                                                                                                  |
+------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
| Sort                                                                                                                                                                                     |
| ├── output columns: [e.name (#1), e.salary (#3), d.department_name (#5), ROW_NUMBER() OVER ( PARTITION BY d.department_id ORDER BY e.salary DESC ) (#6)]                                 |
| ├── sort keys: [department_name ASC NULLS LAST, ROW_NUMBER() OVER ( PARTITION BY d.department_id ORDER BY e.salary DESC ) ASC NULLS LAST]                                                |
| ├── estimated rows: 4.00                                                                                                                                                                 |
| └── Window                                                                                                                                                                               |
|     ├── output columns: [e.name (#1), e.salary (#3), d.department_name (#5), d.department_id (#4), ROW_NUMBER() OVER ( PARTITION BY d.department_id ORDER BY e.salary DESC ) (#6)]       |
|     ├── aggregate function: [row_number]                                                                                                                                                 |
|     ├── partition by: [department_id]                                                                                                                                                    |
|     ├── order by: [salary]                                                                                                                                                               |
|     ├── frame: [Range: Preceding(None) ~ CurrentRow]                                                                                                                                     |
|     └── HashJoin                                                                                                                                                                         |
|         ├── output columns: [e.name (#1), e.salary (#3), d.department_name (#5), d.department_id (#4)]                                                                                   |
|         ├── join type: INNER                                                                                                                                                             |
|         ├── build keys: [d.department_id (#4)]                                                                                                                                           |
|         ├── probe keys: [e.department_id (#2)]                                                                                                                                           |
|         ├── filters: []                                                                                                                                                                  |
|         ├── estimated rows: 8.00                                                                                                                                                         |
|         ├── Exchange(Build)                                                                                                                                                              |
|         │   ├── output columns: [d.department_id (#4), d.department_name (#5)]                                                                                                           |
|         │   ├── exchange type: Merge                                                                                                                                                     |
|         │   └── TableScan                                                                                                                                                                |
|         │       ├── table: default.default.departments                                                                                                                                   |
|         │       ├── output columns: [department_id (#4), department_name (#5)]                                                                                                           |
|         │       ├── read rows: 4                                                                                                                                                         |
|         │       ├── read size: < 1 KiB                                                                                                                                                   |
|         │       ├── partitions total: 1                                                                                                                                                  |
|         │       ├── partitions scanned: 1                                                                                                                                                |
|         │       ├── pruning stats: [segments: <range pruning: 1 to 1>, blocks: <range pruning: 1 to 1>]                                                                                  |
|         │       ├── push downs: [filters: [], limit: NONE]                                                                                                                               |
|         │       └── estimated rows: 4.00                                                                                                                                                 |
|         └── Exchange(Probe)                                                                                                                                                              |
|             ├── output columns: [e.name (#1), e.department_id (#2), e.salary (#3)]                                                                                                       |
|             ├── exchange type: Merge                                                                                                                                                     |
|             └── TableScan                                                                                                                                                                |
|                 ├── table: default.default.employees                                                                                                                                     |
|                 ├── output columns: [name (#1), department_id (#2), salary (#3)]                                                                                                         |
|                 ├── read rows: 10                                                                                                                                                        |
|                 ├── read size: < 1 KiB                                                                                                                                                   |
|                 ├── partitions total: 1                                                                                                                                                  |
|                 ├── partitions scanned: 1                                                                                                                                                |
|                 ├── pruning stats: [segments: <range pruning: 1 to 1>, blocks: <range pruning: 1 to 1>]                                                                                  |
|                 ├── push downs: [filters: [], limit: NONE]                                                                                                                               |
|                 └── estimated rows: 10.00                                                                                                                                                |
+------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
43 rows in set (0.10 sec)
Read 0 rows, 0.00 B in 0.003 sec., 0 rows/sec., 0.00 B/sec.
```

Now
```
+------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
| explain                                                                                                                                                                                  |
+------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
| Sort                                                                                                                                                                                     |
| ├── output columns: [e.name (#1), e.salary (#3), d.department_name (#5), ROW_NUMBER() OVER ( PARTITION BY d.department_id ORDER BY e.salary DESC ) (#6)]                                 |
| ├── sort keys: [department_name ASC NULLS LAST, ROW_NUMBER() OVER ( PARTITION BY d.department_id ORDER BY e.salary DESC ) ASC NULLS LAST]                                                |
| ├── estimated rows: 4.00                                                                                                                                                                 |
| └── Window                                                                                                                                                                               |
|     ├── output columns: [e.name (#1), e.salary (#3), d.department_name (#5), d.department_id (#4), ROW_NUMBER() OVER ( PARTITION BY d.department_id ORDER BY e.salary DESC ) (#6)]       |
|     ├── aggregate function: [row_number]                                                                                                                                                 |
|     ├── partition by: [department_id]                                                                                                                                                    |
|     ├── order by: [salary]                                                                                                                                                               |
|     ├── frame: [Range: Preceding(None) ~ CurrentRow]                                                                                                                                     |
|     └── Exchange                                                                                                                                                                         |
|         ├── output columns: [e.name (#1), e.salary (#3), d.department_name (#5), d.department_id (#4)]                                                                                   |
|         ├── exchange type: Merge                                                                                                                                                         |
|         └── HashJoin                                                                                                                                                                     |
|             ├── output columns: [e.name (#1), e.salary (#3), d.department_name (#5), d.department_id (#4)]                                                                               |
|             ├── join type: INNER                                                                                                                                                         |
|             ├── build keys: [d.department_id (#4)]                                                                                                                                       |
|             ├── probe keys: [e.department_id (#2)]                                                                                                                                       |
|             ├── filters: []                                                                                                                                                              |
|             ├── estimated rows: 8.00                                                                                                                                                     |
|             ├── Exchange(Build)                                                                                                                                                          |
|             │   ├── output columns: [d.department_id (#4), d.department_name (#5)]                                                                                                       |
|             │   ├── exchange type: Broadcast                                                                                                                                             |
|             │   └── TableScan                                                                                                                                                            |
|             │       ├── table: default.default.departments                                                                                                                               |
|             │       ├── output columns: [department_id (#4), department_name (#5)]                                                                                                       |
|             │       ├── read rows: 4                                                                                                                                                     |
|             │       ├── read size: < 1 KiB                                                                                                                                               |
|             │       ├── partitions total: 1                                                                                                                                              |
|             │       ├── partitions scanned: 1                                                                                                                                            |
|             │       ├── pruning stats: [segments: <range pruning: 1 to 1>, blocks: <range pruning: 1 to 1>]                                                                              |
|             │       ├── push downs: [filters: [], limit: NONE]                                                                                                                           |
|             │       └── estimated rows: 4.00                                                                                                                                             |
|             └── TableScan(Probe)                                                                                                                                                         |
|                 ├── table: default.default.employees                                                                                                                                     |
|                 ├── output columns: [name (#1), department_id (#2), salary (#3)]                                                                                                         |
|                 ├── read rows: 10                                                                                                                                                        |
|                 ├── read size: < 1 KiB                                                                                                                                                   |
|                 ├── partitions total: 1                                                                                                                                                  |
|                 ├── partitions scanned: 1                                                                                                                                                |
|                 ├── pruning stats: [segments: <range pruning: 1 to 1>, blocks: <range pruning: 1 to 1>]                                                                                  |
|                 ├── push downs: [filters: [], limit: NONE]                                                                                                                               |
|                 └── estimated rows: 10.00                                                                                                                                                |
+------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
43 rows in set (0.37 sec)
Read 0 rows, 0.00 B in 0.009 sec., 0 rows/sec., 0.00 B/sec.
```

- Fixes #[Link the issue here]

## Tests

- [ ] Unit Test
- [x] Logic Test
- [ ] Benchmark Test
- [ ] No Test  - _Explain why_

## Type of change

- [ ] Bug Fix (non-breaking change which fixes an issue)
- [ ] New Feature (non-breaking change which adds functionality)
- [ ] Breaking Change (fix or feature that could cause existing functionality not to work as expected)
- [ ] Documentation Update
- [ ] Refactoring
- [ ] Performance Improvement
- [x] Other (please describe):

<!-- Reviewable:start -->
- - -
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/datafuselabs/databend/15296)
<!-- Reviewable:end -->